feat: Rewrite to ES Module
BREAKING CHANGE: This project is now exported as an ES Module
