+++ .github/workflows/ci.yml
@@ -33,7 +33,7 @@ jobs:
       matrix:
         os: [ubuntu-latest]
         node-version: ['18', '20', '22']
-        mocha-version: ['6', '7', '8', '9', '10']
+        mocha-version: ['7', '8', '9', '10']
 
     name: 'Test, node: ${{ matrix.node-version }} mocha: ${{ matrix.mocha-version }}'
