+++ lib/mocha-json-runner.mjs
@@ -1,9 +1,9 @@
 /* eslint no-underscore-dangle: ["error", { "allow": ["_afterAll", "_afterEach", "_beforeAll", "_beforeEach"] }] */
 
-const Mocha = require('mocha');
-const mochaJsonDeserialize = require('mocha-json-deserialize');
+import { Runnable, Runner } from 'mocha';
+import mochaJsonDeserialize from 'mocha-json-deserialize';
 
-const { STATE_FAILED, STATE_PASSED } = Mocha.Runnable.constants;
+const { STATE_FAILED, STATE_PASSED } = Runnable.constants;
 const {
   EVENT_RUN_BEGIN,
   EVENT_RUN_END,
@@ -12,7 +12,7 @@ const {
   EVENT_TEST_END,
   EVENT_TEST_PASS,
   EVENT_TEST_PENDING,
-} = Mocha.Runner.constants;
+} = Runner.constants;
 
 function findFailingHooks(hooks) {
   if (hooks && hooks.find)
@@ -28,7 +28,7 @@ const config = { warnOnMissingState: false };
  * @param {string|Object}  json   JSON string or Object
  * @param {boolean}       [delay] Whether or not to delay execution of root suite
  */
-class MochaJsonRunner extends Mocha.Runner {
+class MochaJsonRunner extends Runner {
   constructor(json, delay) {
     let hasStats = false;
     const rootSuite = mochaJsonDeserialize(json);
@@ -165,4 +165,4 @@ class MochaJsonRunner extends Mocha.Runner {
   }
 }
 
-module.exports = MochaJsonRunner;
+export default MochaJsonRunner;
