+++ package.json
@@ -16,34 +16,35 @@
   },
   "license": "MIT",
   "author": "Patrick Seal",
-  "main": "lib/mocha-json-runner.js",
-  "bin": "lib/cli.js",
+  "type": "module",
+  "exports": "./lib/mocha-json-runner.mjs",
+  "bin": "lib/cli.mjs",
   "files": [
     "lib"
   ],
   "scripts": {
-    "coverage-report-lcov": "nyc report --reporter=lcov",
+    "coverage-report-lcov": "c8 report --reporter=lcov",
     "lint": "xo",
     "release": "commit-and-tag-version --sign",
-    "test": "nyc mocha"
+    "test": "c8 mocha"
   },
   "dependencies": {
     "mocha-json-deserialize": "^1.1.1",
-    "yargs": "^17.0.1"
+    "yargs": "^17.7.2"
   },
   "devDependencies": {
-    "chai": "^4.3.4",
+    "c8": "^10.1.2",
+    "chai": "^5.1.1",
     "commit-and-tag-version": "^12.4.1",
+    "es-main": "^1.3.0",
     "mocha": "^10.7.3",
-    "mocha-json-serialize-reporter": "^1.3.0",
-    "nyc": "^15.1.0",
-    "rewire": "^5.0.0",
-    "sinon": "^10.0.0",
-    "test-console": "^1.1.0",
+    "mocha-json-serialize-reporter": "^1.7.0",
+    "sinon": "^18.0.0",
+    "test-console": "^2.0.0",
     "xo": "^0.59.3"
   },
   "peerDependencies": {
-    "mocha": ">=6"
+    "mocha": ">=7"
   },
   "engines": {
     "node": ">=18"
