+++ test/fixtures/mocha-test-simple.fixture.cjs
